Optimize order of provers to reduce proof time
Use of the tool SPAT showed that a lot of time was spend by CVC4 on some VC
that Z3 could prove easily. Changing the order was not a home run though, as
Z3 runs a long time on some VC in sparknack-utils.adb. Solution is to
special-case this unit to run the provers in the other order. On my laptop,
running time goes from 1 mn 50 s to 1 mn 3 s.