Add Really Fast Syndrome Based hash function (RFSB)
This commit adds a first implementation of a version of the RFSB keyed
hash function, which was first introduced in paper referenced at the
end.

This version differs from the one in the paper and in the reference
implementation (SUPERCOP) in the generation of the compressed RFSB-509
matrix. The originally described version of RFSB uses four calls to AES
to generate a single vector while this version uses one call to
ChaCha20. The key used for ChaCha20 is the key for the hash function.

This implemention is not considered to be optimized.

Reference:
Bernstein, D.J., Lange, T., Peters, C., Schwabe, P. (2011). Really Fast
Syndrome-Based Hashing. In: Nitaj, A., Pointcheval, D. (eds) Progress in
Cryptology – AFRICACRYPT 2011. AFRICACRYPT 2011. Lecture Notes in
Computer Science, vol 6737. Springer, Berlin, Heidelberg.
https://doi.org/10.1007/978-3-642-21969-6_9