Add AES256 implementation
This change adds an implementation of the AES256 algorithm. Implemented
are procedures for encrypting/decrypting exactly 16 Bytes in ECB mode.
Those procedures are necessary for the RFSB509 compression function. The
procedures do not process an arbitrary number of bytes as the ECB mode
should not be used for encryption, due to known weaknesses.

The SBox and its inverse are currently not implemented using a constant
time algorithm. This implementation therefore leaks timing information.
An algorithm change is planned.