Remove multiplications in Sbox and Inv_Sbox
The old implementation of Sbox and Inv_Sbox used integer multiplications
with secret data in the Forward_Map and Backward_Map nested procedures.
Integer multiplication is not necessarily a constant time operation on
all CPUs, which creates a side channel attack surface.

This commit subsitutes the multiplications with an algorithm that is a
bit slower but runs in constant time.