Update AES256 calls in RFSB509 procedure/functions
This commit propagates the changes made to the AES256 module to the
RFSB509 module. Beyond the pure syntactic change this commit brings a
performance improvement of about 18% on the SiFive HiFive rev b when
hasing 1024 bytes. The reason for this improvement is that Key_Expansion
function is now called only once per invocation of RFSB509.Hash instead
of four times per hashed byte. This change was not possible before as
the Key_Expansion call was hidden inside Cipher, which was in turn
hidden in ECB_Encrypt.