Use Dynamic_Predicate for Matrix_Column subtype
The Matrix_Column subtype from the RFSB509 module represents elements
from F_2[X]/(x^509 - 1) using 512 bits. This results in the three most
significant bits of the last byte of the subtype to be zero. Before this
commit this invariant was checked using quite some assertions and loop
invariants. Now the Matrix_Column subtype has a dynamic predicate that
assures that the invariant is always preserved.