Add Regression test for RFSB509 procedure
This commit adds a test procedure, that does a regression test of the
RFSB509 procedure against the implementation of the RFSB509 algorithm in
the SUPERCOP toolkit version 20230530. It also adds the fixes that were
needed in order to make the procedure pass the test.

The biggest fix results from the fact, that RFSB509 is specfied to use
AES128 not AES256. Therfore, this commit adds a constant time
implementation of the AES128 algorithm to AES256 module, which was
renamed to AES. For this new implementation tests were added to the test
directory.