Move Broadcast function from AES to Utils module
This commit moves the Broadcast function from the AES to the Utils
module, while renaming the function to Broadcast_Byte. Additionally,
this commit introduces the Broadcast_U16 function in the Utils module,
which sets the upper and lower half of a U32 integer to the value of the
U16 input argument.