Remove Broadcast functions from Utils module
This commit removes the Broadcast_Byte and Broadcast_U16 functions from
the Utils module. It turns out that the functions are not viewed as
static by the compiler thus increasing the cycle count massively. The
Ada2022 version introduces the static aspect, which would help in this
case. But this Repo uses a different Ada version.