Improve Calculate_Column procedure in RFSB509
This commit improves the Calculate_Column procedure in the RFSB509
module. The improvement is that instead of working with the "Vector"
variable and assigning the result to the "Column" variable at the end,
the procedure now directly operates on the "Column" variable. This leads
to less data being copied and therefore a better performance.