Sha384 (#38)
* Added SHA384

* Remove the RR32 and RR64

Both of the mentioned functions simply rename Rotate_Right. To increase
readability this commit removes both functions. Affected modules are
SHA256 and SHA512.

* Rename TS64 to Big_Endian_Unpack

This commit renames TS64 to Big_Endian_Unpack and propagates this change
to the modules implementing SHA256 and SHA512. Additionally,
Big_Endian_Get_Byte ghost function is added to prove the post condition
of the Big_Endian_Unpack.

* Replace TS64 with Big_Endian_Unpack in SHA384 pkg

This commit replaces the TS64 function with the Big_Endian_Unpack
function. Both functions are functionally identical but differ in their
implementation and what is proved about them.

* ReWIP: factored common SHA384, SHA512 code into parent packagage, tests pass still need to run proofs

* SHA384, moved common functions to SHA2_Common pkg

* Workaround for GNAT 14 codegen issue

* Discharged proofs

* Added Loop_Variant to prove subprogram termination

---------

Co-authored-by: mxar <max.arthur@scheven-office.com>